feat: added module security
