+++ admin/src/modules/Security.jsx
@@ -1,0 +1,43 @@
+import { Suspense, lazy } from 'react';
+import { useOutletContext } from 'react-router-dom';
+import { useI18n } from '@wordpress/react-i18n';
+
+import ModuleViewHeader from '../components/ModuleViewHeader';
+import SecurityOverview from '../overview/Security';
+import useModuleSectionRoute from '../hooks/useModuleSectionRoute';
+import { getMapKeysArray } from '../lib/helpers';
+
+const SettingsModule = lazy( () => import( `./static/Settings.jsx` ) );
+
+export default function Security() {
+	const { moduleId } = useOutletContext();
+
+	const tableMenu = new Map( [
+	] );
+
+	const activeSection = useModuleSectionRoute( [
+		'overview',
+		'settings',
+		...getMapKeysArray( tableMenu ),
+	] );
+
+	return (
+		<div className="urlslab-tableView">
+			<ModuleViewHeader
+				moduleId={ moduleId }
+				moduleMenu={ tableMenu }
+				activeSection={ activeSection }
+			/>
+			{
+				activeSection === 'overview' &&
+				<SecurityOverview moduleId={ moduleId } />
+			}
+			{
+				activeSection === 'settings' &&
+				<Suspense>
+					<SettingsModule className="fadeInto" settingId={ moduleId } />
+				</Suspense>
+			}
+		</div>
+	);
+}
